+++ modules/task_1/obolenskiy_a_scalar_product/scalar_product.h
@@ -6,6 +6,6 @@
 #include <string>
 
 std::vector<int> getRandomVector(int sz);
-int getScalarProduct(const std::vector <int> &a, const std::vector <int> &b, size_t count_size_vector);
+int64_t getScalarProduct(const std::vector <int> &a, const std::vector <int> &b, size_t count_size_vector);
 
 #endif  // MODULES_TASK_1_OBOLENSKIY_A_SCALAR_PRODUCT_SCALAR_PRODUCT_H_
